requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2023.1
Django==5.0.2
django-admin-interface==0.28.5
django-cleanup==9.0.0
django-colorfield==0.12.0
django-crispy-forms==2.1
pillow==10.2.0
pypng==0.20220715.0
python-decouple==3.8
python-slugify==8.0.4
qrcode==7.4.2
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
whitenoise==6.6.0
